requirements.txt
autoray==0.6.7
fire==0.5.0
higher==0.2.1
loguru==0.7.2
matplotlib==3.8.2
matplotlib==3.4.3
nbconvert==7.16.3
nbformat==5.10.4
nflows==0.14
numpy==1.23.5
pandas==1.5.3
pyqlib==0.9.3.99
pytz==2023.3.post1
pytz==2022.7.1
PyYAML==6.0.1
PyYAML==6.0.1
ray==2.9.1
scikit_learn==1.3.2
scipy==1.13.0
setuptools==68.0.0
torch==2.0.1
tqdm==4.66.1
